Rename components/CommonSchema to SchemaContainer
